Pulled util functions and indicator defaults into their own files
